avoid infinite loop with bad debuginfo
The immediate failure in PR 16155 is an infinite loop in
value_fetch_lazy.  Each iteration of the loop inside the lval_register
branch computes a value with the same frame id and same register
number as the previous iteration.  It never makes progress, using
progressively more memory creating new values.

It seems to me that it never makes sense to let this loop run
indefinitely.  This patch adds a check and throws an exception if the
same register is returned.  I intentionally did not use an internal
error, because this situation can be caused by bad debuginfo.

I did not go the full distance and have the code check all previous
values.  I could do that if folks want.

With this patch at least the infinite loop is gone.  Now the test case
yields:

    (gdb) bt
    #0  0x0000007fb7ed485c in nanosleep () from /lib64/libc.so.6
    #1  0x0000007fb7ed4508 in sleep () from /lib64/libc.so.6
    #2  0x00000000004008bc in thread_function (arg=0x4) at threadapply.c:73
    #3  0x0000007fb7fad950 in start_thread () from /lib64/libpthread.so.0
    #4  0x0000007fb7f0956c in clone () from /lib64/libc.so.6
    #5  0x0000007fb7f0956c in clone () from /lib64/libc.so.6
    Backtrace stopped: previous frame identical to this frame (corrupt stack?)

2013-11-13  Tom Tromey  <tromey@redhat.com>

	PR backtrace/16155:
	* value.c (value_fetch_lazy): Throw exception if
	get_frame_register_value returns the same register.