accept that value_modifiable makes sense; update comments
Right now the value "modifiable" accessor function is marked as
deprecated.  There's a comment before the declaration that expresses
some skepticism as to the purpose of this flag.  I tracked this
deprecation down to:

    https://www.sourceware.org/ml/gdb-patches/2005-02/msg00026.html

I think this bit makes sense, however.  It is used to ensure that
history values cannot be written to, and in this case one wants to
preserve the original "lval" field -- so a new bit seems appropriate.
This eliminates the "redundant" argument in that email.  As to the
idea of moving it to the piece content, I think this must refer to
some never-implemented plan for evolving values; and in any case it
doesn't seem correct given that this bit applies solely to history
values -- it has nothing to do with pieces.

So, this patch renames the accessor to undeprecate it, and it updates
a couple of comments to make the bit's meaning more clear.

2014-01-09  Tom Tromey  <tromey@redhat.com>

	* ada-lang.c (ada_value_assign, assign_aggregate): Update.
	* breakpoint.c (can_use_hardware_watchpoint): Update.
	* opencl-lang.c (evaluate_subexp_opencl): Update.
	* ppc-linux-nat.c (num_memory_accesses): Update.
	* valops.c (value_assign): Update.
	* value.c (struct value) <modifiable>: Update comment.
	(value_modifiable): Rename from deprecated_value_modifiable.
	* value.h (value_modifiable): Rename from
	deprecated_value_modifiable.