remove deprecated_init_ui_hook
This removes deprecated_init_ui_hook.

The last user (I checked Insight as well) was windows-nat.c, which
used it to install "info dll" as a synonym for "info shared".

I considered implementing a generic way to have a post-initialization
function be run, but in this specific case I think it's reasonable to
simply always have this synonym.  For one thing, one might be doing
remote debugging of a Windows target, and for another, "dll" isn't an
uncommon term in the Unix world anyway.  So, I simply moved the "info
dll" alias to solib.c.

2014-01-09  Tom Tromey  <tromey@redhat.com>

	* defs.h: Remove old comment.
	(deprecated_init_ui_hook): Remove.
	* interps.c (clear_interpreter_hooks): Update.
	* main.c (captured_main): Update comment.
	* solib.c (_initialize_solib): Move "info dll" alias here.
	* top.c (deprecated_init_ui_hook): Remove.
	(gdb_init): Update.
	* windows-nat.c (set_windows_aliases): Remove.
	(_initialize_windows_nat): Remove "info dll" alias.