Remove comment stripping in top.c.
This is not needed as comments are already added to the history (see
the code block just before the removal of the comment stripping in
top.c).  Also when a multi-line command is evaluated as a canned
sequence of commands, GDB tests for comments and returns a nop_command
for that line. GDB continues to interpret lines beginning with "#" as
comments, except jit_control cases where it is passed to the c
compiler (and preprocessor).