compile dwarf expressions to C
This adds code to compile DWARF expressions to C.
Now the generated C function take a pointer to a structure
holding registers from the parent function.
The function may write into this structure; the invoker
has to copy the registers back out afterward.

This patch also has a refactoring in dwarf2-frame.c and dwarf2loc.c
(search for dwarf2_fetch_cfa_info) that should be pulled out as a
separate patch when upstreaming.