make gcc-c-fe.def file
this splits many method definitions out of gcc-interface.h
into a new gcc-c-fe.def file.  this is useful for the plugin.
this also changes a few method signatures and types to make
the api simpler to serialize.

you'll need to make a symlink for the new file into your gcc tree