create a mini AST from the parser
This changes the parser to create a mini AST, that is then walked to
create the gdb expression.  This approach lets us resolve the
reduce/reduce conflict that arises from the need for the parser to
handle both expressions and types.

This may not be sufficient when we add other type productions to the
parser; but it won't hurt.  An alternate plan would be to lex once and
then try the parse two different ways.