Reuse objfiles
FIXME: PR #   also write a test

Sometimes a program consists of a small main executable, but very
large libraries.  For example, most of Firefox is in a single
humongous library, libxul.so.  When re-running the inferior, gdb will
discard the objfiles for the shared libraries (via
objfile_purge_solibs).  But, often gdb will simply need to re-read
them immediately; and in the case described above, this can result in
a noticeable delay.

This patch uses the previous patches to implement an objfile cache.
When a reusable objfile is discarded, it is put on a special list.
Then, the next request for this objfile is satisfied from the cache.

A downside of this approach is that gdb will use more memory.
However, the cache is also purged by "run" -- so objfiles that were
not reused are eventually discarded.  (It would be possible to do this
for an empty "file" command as well, but I did not implement it.)

2018-06-30  Tom Tromey  <tom@tromey.com>

	* symfile.c (symbol_file_add_with_addrs): Use
	make_or_reuse_objfile.
	* objfiles.h (make_or_reuse_objfile): Declare.
	* objfiles.c (preserved_objfiles): New global.
	(zap_preserved_objfiles): New function.
	(free_all_objfiles): Preserve some objfiles.
	(objfile_purge_solibs): Move earlier.  Preserve objfiles in some
	cases.
	(make_or_reuse_objfile): New function.
	(unlink_objfile): Remove assert.