Avoid crashes when stepping through ravenscar context-switching
When stepping through the ravenscar context-switching code, gdb will
try to single-step the current thread, while leaving the other threads
stopped.  This results in an assertion failure in finish_step_over,
because the event thread is not the request thread.

This patch introduces a new target method so that infrun can ask the
target whether this can happen; and then changes infrun to handle this
case.  When the target can have an unexpected thread switch,
single-stepping may also switch threads.

2019-03-26  Tom Tromey  <tromey@adacore.com>

	* infrun.c (finish_step_over): Check
	target_can_randomly_thread_switch.
	(handle_signal_stop): Handle random thread switches.
	* ravenscar-thread.c (struct ravenscar_thread_target)
	<can_randomly_thread_switch>: New method.
	* target-delegates.c: Rebuild.
	* target.h (struct target_ops) <can_randomly_thread_switch>:
	New method.
	(target_can_randomly_thread_switch): New define.