FIXME PR num  - eagerly read debuginfo sections
This changes the DWARF reader to eagerly read the debuginfo sections.

Years ago, I changed dwarf2read.c to lazily read debuginfo sections.
This was done at the time to speed up the reading, on the theory that
some sections were not immediately needed.

However, this lead to at least one bug.  And, because BFD is not
thread-safe, it does not seem safe to read sections in a thread other
than the main thread.

So, this patch removes the lazy reading and instead maps all the
debuginfo sections when creating a DWARF reader.

2019-12-19  Tom Tromey  <tom@tromey.com>

	* dwarf2read.c: Eagerly read debuginfo sections.

Change-Id: I68777ad070c9acaff96831f26e7800743b3ab2aa