Introduce objfile::partial_symtab_future
This adds a new partial_symtab_future member to struct objfile.  When
this is a valid future, it means that partial symbols have not
finished being read -- the future must be "wait"ed upon before partial
symbols can be used.

It also changes some functions in psymtab.c to examine this member.
However, nothing sets it yet.

I looked into making objfile::partial_symtabs a "memoizing future"
that would simply return the partial symtabs.  However, this proved to
be complicated, due to how details of the psymtab API.  It might be
nice to clean this up in the future.

2019-12-15  Tom Tromey  <tom@tromey.com>

	* psymtab.c (done_reading_psymtabs): New function.
	(start_reading_partial_symbols): Check partial_symtab_future.
	Call done_reading_psymtabs.
	(require_partial_symbols): Likewise.
	(psym_has_symbols): Check partial_symtab_future.
	* objfiles.h (struct objfile) <partial_symtab_future>: New
	member.

Change-Id: I9fae8161157fec5b076a842502423fd4c21f0dc3