Read DWARF partial symtabs in worker thread
This changes the DWARF reader to read partial symtabs in a worker
thread.

This is slightly hacky, in that only a single background reader is
allowed.  I don't consider this a big issue.  DWARF is the only
primary supported reader; the only other active one being CTF -- and
if that would benefit from background reading, we could add this.

Similarly, once DWARF is being read in the background for a given
objfile, no other debuginfo format reading can occur for that objfile.
This also is not a major issue.

2019-12-15  Tom Tromey  <tom@tromey.com>

	* dwarf2read.c (dwarf2_do_build_psymtabs): Rename from
	dwarf2_build_psymtabs.
	(dwarf2_build_psymtabs): New function.

Change-Id: I63bf3575e54b8e8f60114ff1eecba04cfbf6f551