Introduce dwarf_psym_reader object
This introduces a new dwarf_psym_reader object.  This is mildly
redundant, given that we already have multiple per-CU objects:
cutu_reader, dwarf2_cu, dwarf2_per_cu_data.  However, this one will
serve a particular role not filled by the others.

In the current patch, the new class is introduced, and various
partial-symbol-related functions are turned into methods on it.  In
the future, a dwarf_psym_reader will be created in one thread, but
then handed to other threads for specialized processing.  The goal is
to have this work be easy to reason about, by having the reader be
relatively self-contained.

2020-02-19  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (struct dwarf_psym_reader): New.
	(process_psymtab_comp_unit_reader): Create a dwarf_psym_reader.
	(build_type_psymtabs_reader): Likewise.
	(dwarf_psym_reader::scan): Rename from scan_partial_symbols.
	Remove parameters.
	(dwarf_psym_reader::add_partial_symbol)
	(dwarf_psym_reader::add_partial_namespace)
	(dwarf_psym_reader::add_partial_module, add_partial_subprogram)
	(dwarf_psym_reader::add_partial_enumeration): Now methods.  Remove
	parameters.