Delay updating the psymtab addrmap
Currently the DWARF partial symbol reader will update the psymtab
addrmap whenever it is convenient.  While this series doesn't
parallelize reading across CUs, in the future it will be useful to do
consolidate these calls, so that we can avoid locking when updating
the addrmap.

In this series, though, the change is still needed, because we are
going to defer building the partial symtab until a bit later; and so
we have to defer the addrmap udpates as well.

2020-02-20  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (struct dwarf_psym_reader) <addresses>: New
	member.
	(process_psymtab_comp_unit_reader): Push addresses on vector.
	Defer call to addrmap_set_empty.
	(dwarf_psym_reader::add_partial_subprogram): Push addresses on
	vector.
	(dwarf2_ranges_read): Change parameter to dwarf_psym_reader.  Push
	addresses on vector.
	(dwarf2_get_pc_bounds): Change parameter to dwarf_psym_reader.