Defer creation of partial symbols
At the end of this series, gdb will have a thread dedicated to simply
creating partial symbols.  It will process each CU as it is read.
This is done to avoid lock contention for the bcache where all
psymbols are stored.

This patch arranges to temporarily store the psymbols in a vector
local to the dwarf_psym_reader, and then introduces a new method to
process the symbols.

This also obsoletes part of commit 96c7f8739 ("[gdb/symtab] Fix
partial unit psymtabs"), and so that is reverted here.

2020-06-20  Tom Tromey  <tom@tromey.com>

	* psymtab.h (class psymtab_storage) <current_global_psymbols,
	current_static_psymbols>: Remove.
	* psymtab.c (partial_symtab::partial_symtab): Don't create new
	vectors.
	(concat): Reove.
	(end_psymtab_common): Simplify.
	(add_psymbol_to_list): Likewise.
	* dwarf2/read.c (struct dwarf_partial_symbol): New.
	(struct dwarf_psym_reader) <symbols>: New member.
	<load_partial_dies, do_add_psymbol_to_list, create_symbols>:
	New methods.
	(process_psymtab_comp_unit_reader, build_type_psymtabs_reader)
	(load_partial_comp_unit): Update.
	(dwarf_psym_reader::create_symbols): New method.
	(process_psymtab_comp_unit_reader, build_type_psymtabs_reader)
	(load_partial_comp_unit): Update.
	(dwarf_psym_reader::do_add_psymbol_to_list): New method.
	(dwarf_psym_reader::add_partial_symbol): Use
	do_add_psymbol_to_list.
	(dwarf_psym_reader::load_partial_dies): Rename from
	load_partial_dies.  Use do_add_psymbol_to_list.