Split creation of include psymtabs into two parts
Following the theme of deferring the creation of objects directly on
the partial symtab, this splits the handling of the include psytabs
into two parts.  First, we compute the data and store it on the
dwarf_psym_reader.  Then, later, we create the include psymtabs

The goal of this is to avoid locking when creating the psymtab, by
handling all psymtab creation in a single thread.

2020-02-20  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (dwarf2_build_include_psymtabs): Add "dirname"
	and "psym_reader" parameters.  Remove "pst" parameter.
	(struct dwarf_psym_reader) <include_names, name_storage, filename,
	dirname>: New members.
	(process_psymtab_comp_unit_reader): Set new members.  Call
	dwarf2_build_include_psymtabs earlier.  Create include psymtabs at
	end.
	(psymtab_include_file_name): Change "pst" parameter to
	"psym_reader".
	(dwarf_decode_lines): Likewise.