Do not copy names onto storage obstack
Because obstacks are not thread-safe, we have to take some care about
when to use them.  In particular, in the threading patch at the end of
this series, the psymtab-reading thread may use the per-BFD storage
obstack.  However, names may also be allocated there.

To avoid locking (and thus contention), we instead arrange to
temporarily put such names onto the comp_unit_obstack.  This is ok
because the worker thread may touch the storage obstack or the bcache
-- but not the comp_unit_obstack.

We have to make an intermediate copy of these names on the heap,
because the comp_unit_obstack may already have been destroyed by the
time the symbol is created.  This works out ok, though, because other
code may also allocate names on the heap, and so the temporary symbol
structure we're using already accounts for this.

This patch saves some memory, because now the names end up in a
bcache.  Previously, each time a name was seen, it would be copied to
the storage obstack instead.

For "file ./gdb", my numbers went from:

Space used: 33054720 (+24719360 for this command)

to:

Space used: 18321408 (+9957376 for this command)

... which seems rather large now that I look at it.

2020-02-19  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (struct partial_die_info) <copy_name>: New
	member.
	(partial_die_info): Initialize copy_name.
	(partial_die_full_name): Copy name if needed.
	(dwarf_psym_reader::load_partial_dies)
	(dwarf_psym_reader::load_partial_dies): Copy name if needed.
	(partial_die_info::read, guess_partial_die_structure_name)
	(partial_die_info::fixup): Use comp_unit_obstack.  Set copy_name.