Introduce dwarf2_unshareable and move die_type_hash
The goal of this series is to share dwarf2_per_objfile across
objfiles.  However, this is difficult to do, because it has some
members that are inherently per-objfile.

This patch introduces a new "unshareable" object.  The idea here is
that we'll have one such object per objfile, and we'll temporarily
update the pointer when needed.  This is a bit ugly, but it's
convenient, because it doesn't require a huge rewriting of the DWARF
code.  (I would be in favor of such a rewriting, but ... it's huge.)

This patch further moves one such member, the DIE type hash, to the
new structure.

In this patch, the new object is managed by the dwarf2_per_objfile.
However, by the end of the series, this will change.

gdb/ChangeLog
2020-02-21  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.h (struct dwarf2_unshareable): New type.
	(struct dwarf2_per_objfile) <die_type_hash>: Move to
	dwarf2_unshareable.
	<unshareable>: New member.
	* dwarf2/read.c (dwarf2_per_objfile::dwarf2_per_objfile):
	Initialize "unshareable".
	(dwarf2_per_objfile::~dwarf2_per_objfile): Delete unshareable.
	(set_die_type, get_die_type_at_offset): Update.