Don't intern the filename in partial_symtab constructor
Right now, the partial_symtab constructor interns the file name in the
filename cache.

However, in the DWARF reader, we plan to have all uses of the filename
cache be done in one thread, while creation of partial symtabs will be
done in another thread.

This patch moves the interning to the standard_psymtab subclass (so
other symbol readers continue to get this behavior), and changes the
DWARF reader to do the interning on its own.

2020-02-22  Tom Tromey  <tom@tromey.com>

	* objfiles.h (struct objfile) <intern>: New method.
	* dwarf2/read.c (dwarf2_create_include_psymtab): Intern the name.
	(create_partial_symtab): Likewise.
	* psympriv.h (struct standard_psymtab) <standard_psymtab>: Intern
	filename.
	* psymtab.c (partial_symtab::partial_symtab): Don't intern the
	filename.