Fix filename interning for ordinary symtabs
An earlier patch changed the DWARF reader to intern file names for
include partial symtabs in dwarf_psym_reader::intern_names.  This
patch does the same for ordinary partial symtabs.  The idea here is
that all access to the string bcache will be done in one spot: the
intern_names method.  Later, we'll be able to invoke this function in
a separate thread and not worry about locking access to the filename
cache.

2020-02-22  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (create_type_unit_group): Intern "name".
	(create_partial_symtab): Don't intern "name".
	(struct dwarf_psym_reader) <interned_filename>: New member.
	(process_psymtab_comp_unit_reader): Use interned_filename.
	(dwarf_psym_reader::intern_names): Set interned_filename.