Simple speedup for DWARF CU expansion
I noticed that DWARF CU expansion had an easy-to-fix hot spot:
following DIE references in dwarf2_attr.

This patch fixes the problem by caching the referent in the die_info,
if the two DIEs are in the same CU.  (This restriction avoids any
issues with CU invalidation.)

I ran "gdb -readnow" on itself 10 times.  The mean before on this
machine was 14.522 seconds; with the patch it was 13.685 seconds.

Historically I figured that CU expansion was not very important; but I
think occasionally it results in unexpected delays for users.  So,
it's probably worth fixing.  I'm not sure the micro-optimization
approach is really all that good in the long run.  Better, I think,
would be to read types and function bodies lazily -- this would vastly
improve performance.  (I attempted the latter once and saw a 40%
speedup, IIRC.)

The use of dwarf2_attr could also be improved.  Rather than looking up
attributes individually, it would probably be better for (most)
functions to loop once over all attributes, collecting useful
information.

Meanwhile, this patch seems reasonably worthwhile.

gdb/ChangeLog
2020-05-23  Tom Tromey  <tom@tromey.com>

	* dwarf2/read.c (dwarf2_attr): Use die_info::ref.
	* dwarf2/die.h (struct die_info) <ref>: New member.