Use ada-operator.def to define operator names
There's a comment in ada-lang.c about keeping attribute_names
up-to-date with respect to ada-lang.h.  This is actually incorrect and
it should refer to ada-operator.def instead.  However, keeping things
aligned like this is error-prone, and so I came up with this patch to
make it more difficult to introduce an error here.

gdb/ChangeLog
2020-08-05  Tom Tromey  <tromey@adacore.com>

	* expression.h (XOP): New define.
	* ada-lang.c (attribute_names): Use ada-operator.def to fill
	array.
	* ada-operator.def: Use XOP for attribute operators.