Set first_die_cu_offset in read_comp_unit_head
Currently, read_and_check_comp_unit_head sets the first_die_cu_offset
member, but read_comp_unit_head does not.  This seems a bit strange,
because this means that the value isn't always guaranteed to be
correct.  This patch arranges to set it in read_comp_unit_head, so
that it can be used unconditionally.

gdb/ChangeLog
2021-05-11  Tom Tromey  <tom@tromey.com>

	* dwarf2/comp-unit.c (read_comp_unit_head): Set
	first_die_cu_offset.
	(read_and_check_comp_unit_head): Don't set first_die_cu_offset.