Add zebra-striping to CLI table display
This patch adds two new style settings to gdb.  These can be used to
add "zebra striping" to CLI tables -- that is, change the prevailing
style of alternating lines of a table.

Modifying the CLI ui-out is sufficient for most tables.  However, this
did not work for stack traces, because those aren't ordinary tables.
This patch uses the "phony table" support (added earlier in this
series) to achieve the same effect without otherwise impacting stack
trace display.

By default, styles are only changed if 256 colors are available.  This
way we can choose a relatively innocuous grey background.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=32285